fix: file moved into DTO
